Disambiguate __ffsll in the volume rebuild path
HIP declares __ffsll for both long long and unsigned long long, so
passing a uint64_t matches neither exactly:

    volume_builder.h:249: error: call to '__ffsll' is ambiguous

CUDA declares a single overload, so the bare call compiles there and
this only surfaces on HIP. Cast explicitly rather than changing the
parameter type, which would ripple through the callers.

Reachable only because __CUDA_ARCH__ is now defined: the #else branch is
a portable bit-scan loop that had been selected on HIP until this point.

+++ warp/native/volume_builder.h
@@ -246,7 +246,11 @@ CUDA_CALLABLE inline uint32_t rebuild_find_lowest_on(uint64_t word)
         return 64u;
     }
 #if defined(__CUDA_ARCH__)
-    return uint32_t(__ffsll(word) - 1);
+    // Cast explicitly: HIP declares __ffsll for both long long and unsigned
+    // long long, so a uint64_t argument matches neither exactly and the call is
+    // ambiguous. CUDA declares a single overload, so the bare call compiles
+    // there and this only surfaces on HIP.
+    return uint32_t(__ffsll(static_cast<unsigned long long>(word)) - 1);
 #else
     for (uint32_t i = 0; i < 64u; ++i) {
         if (word & (uint64_t(1) << i)) {
